feat(mail): organize settings into tabs
